// lib/api/generated/models/BusinessEstimateResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BusinessEstimateResponse = {
    estimated_total: number;
    sample_results: number;
    cost_estimate: number;
    location: string;
    query: string;
};