// lib/api/generated/models/FreeEstimateResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FreeEstimateResponse = {
    estimated_total: number;
    business_density: string;
    location_info: Record<string, any>;
    query: string;
    cost: number;
};