// lib/api/generated/models/LeadCreditCheckout.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Serializer for lead credit checkout
 */
export type LeadCreditCheckout = {
    /**
     * Stripe price ID for the credit package
     */
    price_id: string;
    readonly url: string;
};